permissions

+++ app/Http/Controllers/EventController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use App\Http\Requests\EventStoreRequest;
+use Carbon\Carbon;
 
 class EventController extends Controller
 {
@@ -19,7 +20,7 @@ class EventController extends Controller
     }
 
     public function list($userId, Request $request) {
-        $eventlist = Event::all();
+        $eventlist = Event::whereDate('starttime', '>=', Carbon::today())->get();
 
         foreach($eventlist as $event)
         {
@@ -47,15 +48,18 @@ class EventController extends Controller
         $event = Event::findOrFail($eventId);
         $participating = count($event->users()->where('user_id', $userId)->get());
 
-        if ($participating == 0) {
-            $event->users()->attach($userId);
-            $enrolled = 1;
-        } elseif ($participating > 0) {
-            $event->users()->detach($userId);
-            $enrolled = 0;
+        if ($event['enrolltime'] >= Carbon::now()){
+            if ($participating == 0) {
+                $event->users()->attach($userId);
+                $enrolled = 1;
+            } elseif ($participating > 0) {
+                $event->users()->detach($userId);
+                $enrolled = 0;
+            }
+        } else {
+            return response()->json(['message'=>'Enrollment is closed'], 406);
         }
-        
-        $arr = json_decode($event, true);
+                $arr = json_decode($event, true);
         $arr['enrolled']=$enrolled;
         $arr['participants'] = count($event->users()->get());
         $event = json_encode($arr);
